Gave functionality to banner on hero section
